requirements.txt
torch==2.0.0
torchvision==0.15.1
clip-by-openai==1.0
wandb==0.15.0
fastapi==0.89.1
uvicorn==0.21.1
python-dotenv==0.21.1
pillow==9.3.0
numpy==1.23.4
pandas==1.5.3
tqdm==4.64.1
requests==2.28.1
deep-translator==1.7.3
gunicorn==20.1.0
cloudinary==1.29.0